+++ lua/plugins/persistence.lua
@@ -1,0 +1,9 @@
+-- lua/plugins/persistence.lua
+
+return {
+    'folke/persistence.nvim',
+    event  = 'BufReadPre',
+    config = function()
+        require('config.persistence')
+    end,
+}